+++ example/a.ts
@@ -1,19 +1,10 @@
-import { Elysia, t } from '../src'
-
-const T = t.Union([t.Literal('order'), t.Literal('artists')])
+import { Elysia } from '../src'
 
 const app = new Elysia()
-	.onBeforeHandle(({ headers }) => {
-		console.log(headers)
-	})
-	.all('/v3/:id', (ctx) => {
-		return 'Hello from http'
-	})
-	.ws('/v3/:id', {
-		message(ws, message) {
-			ws.send('pong')
-		}
+	.get('/', ({ cookie: { access_token, refresh_token } }) => {
+		access_token.value = 'none'
+		refresh_token.value = 'none'
+
+		return 'hi'
 	})
 	.listen(3000)
-
-console.log(app.fetch.toString())